# app/DTO/UserDTO.php
<?php

namespace App\DTO;

use App\Models\User;

final class UserDTO
{
    public int $id;
    public string $name;
    public string $email;

    public function __construct(int $id, string $name, string $email)
    {
        $this->id = $id;
        $this->name = $name;
        $this->email = $email;
    }

    public static function fromModel(User $user): self
    {
        return new self($user->id, $user->name, $user->email);
    }

    public function toArray(): array
    {
        return [
            'id' => $this->id,
            'name' => $this->name,
            'email' => $this->email,
        ];
    }
}